+++ configfile.js
@@ -29,7 +29,8 @@ cfreader._enoent_files = {};
 cfreader._sedation_timers = {};
 cfreader._overrides = {};
 
-const config_dir_candidates = [
+let config_dir_candidates = [
+    // these work when this file is required as './config.js'
     path.join(__dirname, 'config'),    // Haraka ./config dir
     __dirname,                         // npm packaged plugins
 ];
@@ -47,10 +48,15 @@ function get_path_to_config_dir () {
         return;
     }
 
+    // these work when this file is required as 'haraka-config'
     if (/node_modules\/haraka-config$/.test(__dirname)) {
-        // loaded by a npm packaged module
-        cfreader.config_path = path.resolve(__dirname, '..', '..');
-        return;
+        config_dir_candidates = [
+            // haraka/Haraka/tests/*
+            path.resolve(__dirname, '..', '..', 'config'),
+
+            // npm packaged modules
+            path.resolve(__dirname, '..', '..'),
+        ]
     }
 
     for (let i=0; i < config_dir_candidates.length; i++) {